feat(api): implement ticket management API with authentication
- Added TicketApiController for handling ticket-related API requests.
- Updated routes to use TicketApiController with middleware for API key access.
- Implemented CRUD operations for tickets including listing, creating, updating, closing, and deleting tickets.
- Introduced EnsureAccessApiKey middleware to validate API key for incoming requests.
- Created comprehensive API documentation for ticket management.
- Added pnpm workspace configuration for build management.

+++ app/DTOs/Ticket/StoreTicketDto.php
@@ -34,7 +34,7 @@ class StoreTicketDto
         return new self(
             title: $data['title'],
             description: $data['description'],
-            requesterId: $data['requester_id'] ?? Auth::id(),
+            requesterId: (int) ($data['requester_id'] ?? Auth::id()),
             images: $data['images'] ?? null,
             type: $data['type'],
             impact: $data['impact'],
